requirements.txt
# Discord
discord.py>=2.3.0

# Async HTTP
aiohttp>=3.9.0

# Async SQLite
aiosqlite>=0.19.0

# Config parsing
pyyaml>=6.0